rag-service/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
markdownify==0.13.1
sentence-transformers==3.0.1
faiss-cpu==1.12.0
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.6
overpy==0.7
python-dotenv==1.0.1
colorama==0.4.6
rich==13.8.1
